requirements.txt
asgiref==3.5.2
attrs==21.4.0
click==8.1.3
fastapi==0.68.2
h11==0.13.0
iniconfig==1.1.1
mysql-connector-python==8.0.29
packaging==21.3
pluggy==1.0.0
protobuf==4.21.1
py==1.11.0
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
starlette==0.14.2
tomli==2.0.1
typing_extensions==4.2.0
uvicorn==0.15.0